Audiobookshelf Integration Migration (partial)
