+++ crates/services/integration/src/integration_type.rs
@@ -1,8 +1,10 @@
 use enums::MediaSource;
+use providers::google_books::GoogleBooksService;
 
 pub enum IntegrationType {
     Komga(String, String, String, MediaSource),
     Jellyfin(String),
     Emby(String),
     Plex(String, Option<String>),
+    Audiobookshelf(String, String, GoogleBooksService)
 }
