+++ crates/services/integration/src/lib.rs
@@ -1,12 +1,7 @@
-use std::future::Future;
-
-use anyhow::{anyhow, bail, Result};
-use application_utils::get_base_http_client;
-use async_graphql::Result as GqlResult;
+use anyhow::{bail, Result};
 use database_models::metadata;
 use enums::{MediaLot, MediaSource};
-use media_models::{CommitMediaInput, IntegrationMediaCollection, IntegrationMediaSeen};
-use providers::google_books::GoogleBooksService;
+use media_models::{IntegrationMediaCollection, IntegrationMediaSeen};
 use radarr_api_rs::{
     apis::{
         configuration::{ApiKey as RadarrApiKey, Configuration as RadarrConfiguration},
@@ -14,8 +9,6 @@ use radarr_api_rs::{
     },
     models::{AddMovieOptions as RadarrAddMovieOptions, MovieResource as RadarrMovieResource},
 };
-use reqwest::header::{HeaderValue, AUTHORIZATION};
-use rust_decimal_macros::dec;
 use sea_orm::DatabaseConnection;
 use sonarr_api_rs::{
     apis::{
@@ -24,7 +17,6 @@ use sonarr_api_rs::{
     },
     models::{AddSeriesOptions as SonarrAddSeriesOptions, SeriesResource as SonarrSeriesResource},
 };
-use specific_models::audiobookshelf as audiobookshelf_models;
 use traits::TraceOk;
 
 use crate::{
@@ -32,6 +24,7 @@ use crate::{
     integration_type::IntegrationType,
     komga::KomgaIntegration
 };
+use crate::audiobookshelf::AudiobookshelfIntegration;
 use crate::emby::EmbyIntegration;
 use crate::jellyfin::JellyfinIntegration;
 use crate::plex::PlexIntegration;
@@ -44,6 +37,7 @@ mod jellyfin;
 mod emby;
 mod show_identifier;
 mod plex;
+mod audiobookshelf;
 
 #[derive(Debug)]
 pub struct IntegrationService {
@@ -65,139 +59,7 @@ impl IntegrationService {
         Ok(payload)
     }
 
-    pub async fn audiobookshelf_progress<F>(
-        &self,
-        base_url: &str,
-        access_token: &str,
-        isbn_service: &GoogleBooksService,
-        commit_metadata: impl Fn(CommitMediaInput) -> F,
-    ) -> Result<(Vec<IntegrationMediaSeen>, Vec<IntegrationMediaCollection>)>
-    where
-        F: Future<Output = GqlResult<metadata::Model>>,
-    {
-        let client = get_base_http_client(
-            &format!("{}/api/", base_url),
-            Some(vec![(
-                AUTHORIZATION,
-                HeaderValue::from_str(&format!("Bearer {access_token}")).unwrap(),
-            )]),
-        );
-        let resp = client
-            .get("me/items-in-progress")
-            .send()
-            .await
-            .map_err(|e| anyhow!(e))?
-            .json::<audiobookshelf_models::Response>()
-            .await
-            .unwrap();
-        tracing::debug!("Got response for items in progress {:?}", resp);
-        let mut media_items = vec![];
-        for item in resp.library_items.iter() {
-            let metadata = item.media.clone().unwrap().metadata;
-            let (progress_id, identifier, lot, source, podcast_episode_number) =
-                if Some("epub".to_string()) == item.media.as_ref().unwrap().ebook_format {
-                    match &metadata.isbn {
-                        Some(isbn) => match isbn_service.id_from_isbn(isbn).await {
-                            Some(id) => (
-                                item.id.clone(),
-                                id,
-                                MediaLot::Book,
-                                MediaSource::GoogleBooks,
-                                None,
-                            ),
-                            _ => {
-                                tracing::debug!("No Google Books ID found for ISBN {:#?}", isbn);
-                                continue;
-                            }
-                        },
-                        _ => {
-                            tracing::debug!("No ISBN found for item {:#?}", item);
-                            continue;
-                        }
-                    }
-                } else if let Some(asin) = metadata.asin.clone() {
-                    (
-                        item.id.clone(),
-                        asin,
-                        MediaLot::AudioBook,
-                        MediaSource::Audible,
-                        None,
-                    )
-                } else if let Some(itunes_id) = metadata.itunes_id.clone() {
-                    match &item.recent_episode {
-                        Some(pe) => {
-                            let lot = MediaLot::Podcast;
-                            let source = MediaSource::Itunes;
-                            let podcast = commit_metadata(CommitMediaInput {
-                                identifier: itunes_id.clone(),
-                                lot,
-                                source,
-                                ..Default::default()
-                            })
-                            .await
-                            .unwrap();
-                            match podcast
-                                .podcast_specifics
-                                .and_then(|p| p.episode_by_name(&pe.title))
-                            {
-                                Some(episode) => (
-                                    format!("{}/{}", item.id, pe.id),
-                                    itunes_id,
-                                    lot,
-                                    source,
-                                    Some(episode),
-                                ),
-                                _ => {
-                                    tracing::debug!(
-                                        "No podcast found for iTunes ID {:#?}",
-                                        itunes_id
-                                    );
-                                    continue;
-                                }
-                            }
-                        }
-                        _ => {
-                            tracing::debug!("No recent episode found for item {:#?}", item);
-                            continue;
-                        }
-                    }
-                } else {
-                    tracing::debug!("No ASIN, ISBN or iTunes ID found for item {:#?}", item);
-                    continue;
-                };
-            match client
-                .get(format!("me/progress/{}", progress_id))
-                .send()
-                .await
-                .map_err(|e| anyhow!(e))?
-                .json::<audiobookshelf_models::ItemProgress>()
-                .await
-            {
-                Ok(resp) => {
-                    tracing::debug!("Got response for individual item progress {:?}", resp);
-                    let progress = if let Some(ebook_progress) = resp.ebook_progress {
-                        ebook_progress
-                    } else {
-                        resp.progress
-                    };
-                    media_items.push(IntegrationMediaSeen {
-                        lot,
-                        source,
-                        identifier,
-                        podcast_episode_number,
-                        progress: progress * dec!(100),
-                        provider_watched_on: Some("Audiobookshelf".to_string()),
-                        ..Default::default()
-                    });
-                }
-                _ => {
-                    tracing::debug!("No progress found for item {:?}", item);
-                    continue;
-                }
-            };
-        }
-        Ok((media_items, vec![]))
-    }
+
 
     pub async fn radarr_push(
         &self,
@@ -277,6 +139,10 @@ impl IntegrationService {
             IntegrationType::Plex(payload, plex_user) => {
                 let plex = PlexIntegration::new(payload, plex_user, self.db.clone());
                 plex.progress().await
+            }
+            IntegrationType::Audiobookshelf(base_url, access_token, isbn_service) => {
+                let audiobookshelf = AudiobookshelfIntegration::new(base_url, access_token, isbn_service);
+                audiobookshelf.progress().await
             }
         }
     }
